CHANGE STRUCTURE OF THE WEBPAGE.
I have changed the structure of the webpage to have the background image and the form side by side by positioning them. I have also added the logo to the heading and further positioned both to be lower on the page. I have also designed and postioned the footer to the bottom of the webpage.