CHANGE THE DIV BEHIND THE HEADING AND LOGO.
I have styled the div behind the logo and heading by positioning it and giving it a height to align with the logo and heading. I have also added a background color of black and an opacity of 0.65 to be able to clearly read the heading against the busy background.